Get prototype templates into cms app
The pages are viewable under url /site

Much the public facing site is in the directory
static_site and has it's own static and template directories.

JS and SASS for the public sites re in src directory.

Tasks avaialable to build js and sass are:

gulp sass
gulp watch

Any data loaded into the pages are in static_site/data
however the goal is to load data from rd_content repo. That is
therefore what we need to do asap.