Update accordion javascript
https://trello.com/c/qf7kEvxS/470-accordions-should-be-open-without-java
script-and-theres-no-show-hide-link

When javascript isn’t available, all of the sections should be
expanded, with no buttons or section dividers, rather than permanently
collapsed, as is currently the case.

I ended up re-writing most of the javascript to make it work via
progressive enhancement, and also adding the correct `aria-expanded`
attributes for accessibility.