Ethnic groups in England and Wales by economic status
See
https://trello.com/c/MGoEa8Ku/427-ethnic-groups-in-england-and-wales-by-
economic-status-static-page

I’ve copied all the content across from the Google doc. Unfortunately
the screenshot of the graph is super low-res, and the table screenshots
are unusable (as they don’t show the full table). These have all been
removed from the training CMS and need to be re-created.