Listing tables explicitly in pg_dump in a specific order, means that when we restore we avoid foreign key constraint errors.
Disabling triggers was partially sucessful in that we did still
load data corectly. However as the disable triggers flag also
tried to disable postgres system triggers (which failed as
the postgres user is not a superuser) the result was a non 0 exit
code that confused heroku and git about what the outcome of the deploy
was.

This avoids a non zero exit code but allows us to load realistic
data into preview apps.