The purpose of this pr is to remove commited css and js from repository and replace with versioned assets that can be built when needed on dev and deployed code.
Also removed some redundant files:
main.css not used
measure.js empty

Moved all vendor code from src, it's not our code
and in any case already existed in static. No point
in watching in gulp and copying over. Likewise gov.uk js
can just live in static

The files in src are actually our source files, that we
work on and get concatenated into all-*.js

Not done in this PR:

- Move of chart/table js to src and therefore into all-*.js bundle.
The code as it is now is not, out of the box able to be just
added to the bundle. May need work to make into modules?

- Some pages in cms still have in page js.

Both of the above need more work and shaping up if possible.

If this gets merged into master, when you pull into local dev you'll
need to run

```
gulp sass scripts
```

to get versioned assets, and if you're working on sass/js you should be running

```
 gulp watch
```

Downside of this change, if you modify js or sass and get new versions
flask dev server will not be aware of the changes and you will not get
reload for free which is a pain (jinja filters are cached) so with
SO's help I added a small before request handler that turns off
jinja caching in dev.