Add `package-lock.json`
I upgraded to NPM `5.4.2` locally to fix an issue with the `gulp-rev`
module, which also means this `package-lock.json` is generated, to
document the exact versions of all the dependencies and
sub-dependencies used. In theory this should help to solve issues with
different npm modules being used in development vs production.