Heroku deploys are now quite slow as we force a build on a deploy.
This pr change that to a request for a build that leaves the actual
work of doing the build to the scheduled job to pick up.

Therefore on deploy a build of static site may be up to ten minutes
later, but the deploy of the heroku app will not be held up while
a forced static build runs.

There is still a management command to force the build.

I'm not sure if it makes sense to still have the remaining build
static task, but that's another question.