On upload try to decode as utf-8 and then on download return as utf-8.
If unicode decode error encountered, try to decode as iso-8859-1
and return it in same encoding.

The above will handle anything that is plain ascii just fine.

I've had a look at lots of csv downloads on gov.uk, and those that
include the £ symbol (the thing that alerted us to the problem) are
encoded as iso-8859-1. In fact most files I found are encoded
as iso-8859-1 with a small number encoded as ascii.