Detect if radio/checkboxes selected
When the page loads, we need to check whether to display the secondary
source sections or not, based on whether there is any data specified
for any of those sources.  The current code just checks the `value`
attribute of all the `<input>` elements within the fieldset. For radio
and checkboxes we should detect whether any of them were `checked` or
not instead.

Also, upon removing a secondary source, we should set `checked` to
`false` to reset the radio and checkboxes back to their default state.