Use the first bullet point as the meta description
This specifies the first bullet point as the ‘description’ in the page
metadata, which is what gets used on Twitter, Facebook and other
platforms as a preview snippet.

Not specifying a meta description means that these platforms sometimes
select their own sentences from the page to use as a preview, which is
less ideal (eg as they sometimes select the first paragraph of ‘things
you need to know’).