Ensure static files are uploaded to S3 before updated pages
We've seen an issue where new/updated pages are uploaded before the CSS they need, and so users can see the pages without CSS.  This persists for 15 minutes because of cacheing, even if the required files upload in the meantime.

By ensuring all static assets are uploaded first we should prevent this from happening again.