Update CMS views and forms to use new Contact 2 fields
We now have contact details per page, not per source.  This updates CMS
pages to use the new fields and not the old fields which have been
deleted.

This also fixes a bug in the export template that had some content
wrapped in multiple two-thirds columns, leading to very narrow text
on the page.