BUGFIX: Wrong "topic" in links to new measures from dashboard
Loinks for new topics were being generated with two duplicate slugs in
the path (page.parent.uri).  This worked OK in the CMS for some reason,
so no-one noticed the bad URLs.  But when adding the dashboards to the
static build these double-slugged links no longer work, and they need
the higher level categorey to be present in the path (ie
page.parent.parent.uri).