BUGFIX: Save "Other" publication frequency field on first try
When deciding whether to update the `page.frequency_other` field we were
previously looking up `page.frequency_of_release.description` to check
if it was "Other".  However, this is a relationship to the
FrequencyOfRelease table that would only be updated once the page had
finished saving, so was referring to the *previous* description.

By changing this to lookup the latest value from FrequencyOfRelease
directly this will now recognise when the "Other" radio button has been
selected and so save the `frequency_other` value as intended.

Also removed the check for `page.frequency_of_release is not None`
because again this is referring to the *previous* state of the object
and will block saving the first time (as page.frequency_of_release
will be None until after the update has saved).

Also removed the check for `frequency_other is not None` to reduce the
complexity of the conditional, as it makes no difference either way -
if the conditional here fails because frquency_other is None then
the fall-through is just to save None anyway.