Require file for uploads form
Previously, if the form was submitted without a file having been chosen
then an error was thrown by the upload_service as it tried to get a
filename from the empty upload field.

By adding validation at the form level that a file has been chosen we
can avoid the error and instead show a validation message on the upload
field.