Use Jinja lstrip and trim instead of BeautifulSoup to reduce whitespace
We were using BeautifulSoup to "prettify" our HTML in the static site
build, but it has a tendency to add whitespace at the end of inline
links which shouldn't be there.

Jinja's `trim_blocks` and `lstrip_blocks` directives can be used to
control the amount of whitespace in rendered templates (see
http://jinja.pocoo.org/docs/dev/templates/#whitespace-control).

The results look good, and we don't get those pesky extra spaces in the
links.

Build with no processing: 85.3 MB
Current BeautifulSoup build: 80.2 MB
New Jinja stripped build: 75.8 MB