Add favicon(s) to CMS base template
The recently published dashboard pages don't have a favicon because they
use the CMS base template, which for some reason differs from the static
site base template.

It would probably be nice to fix it so the CMS and static site both use
the same base template, but for now I think we can just add the icons to
the CMS template.