Use `main_content` as block name everywhere
I noticed that error pages in the CMS don't render any content on the
page - it's just a header and footer. This is because the error page
templates have `block content` but the CMS base template includes only
`block main_content`.

There was a workaround in the export templates where they wrapped
`block main_content` inside `block content`.

It feels nicer to just use the same name everywhere.