BUGFIX: `date_first_published` is not a field; `published_date` is
The "Date first published" section of the primary data source was never
being shown as the field referenced in the template does not exist in
the model.

It turns out that we don't want to show these dates just yet as the data
might not be correct or consistent, so existing behaviour is maintained
here.  But it is now by design rather than by chance that it does what
we want.