Re-order view decorators so unauth users redirect to login not Forbidden
Previously if a user wasn't logged in and tried to access a CMS page
also wrapped with `user_can` or `user_has_access` they would get a 403
forbidden rather than be redirected to the login page.

By re-ordering the decorators so that `login_required` is the outermost
decorator users who are not logge din will be redirected to the login
page rather than flat denied.