Remove unused page fields.
Moved sql to manage views to external file.

This is precursor step to a rebase of migrations.

The actual rebase will be in next pr and name of
rebase migration will be be same as migration in this
pr and will therefore not actually be applied.

The point that is it could be used to start fresh environment
and be starting point of subsequent migrations.