Link to survey at top of all pages
This updates the 'BETA' banner at the top of every page to include a call to action to answer our short survey.

The banner is reversed-out (white on blue) for increased prominance.

We can switch between this banner and the existing one using an envirnment variable, so that this is easy to turn on and off.