CMS views check topic/subtopic/measure hierarchy is correct
Previously some of the CMS views didn't enforce that the topic/subtopic
mentioned in the URL are the ones that the measure belongs to.

It's a bit odd that you could update a measure at a URL that has
mismatched topic and subtopic parameters, so this makes CMS endpoints
that take topic/subtopic/measure/version check that the hierarchy
is consistent.