BUGFIX: Handle empty 'title' object for some chart types
In the Chartbuilder 1 form there are fields for xAxis and yAxis titles.
Because these submit as empty strings with the form even if nothing is
typed in to them most charts created with Chartbuilder 1 have data that
looks like:
```
{"xAxis": {"title":{"text": ""}, ...}},
{"yAxis": {"title":{"text": ""}, ...}},
```
ie there is always a "text" key inside the title object, even if it is
mostly empty.

In the new streamlined Chartbuilder 2 form there are no boxes for x and
y axis titles (as we don't show them in the charts anyway). This means
that the chart objects created with Chartbuilder 2 have data that looks
like:
```
{"xAxis": {"title":{}, ...}},
{"yAxis": {"title":{}, ...}},
```

This causes problems with our data downloads because the code currently
expects there to always be a `text` key inside the `title` object.

After some deliberation about which way to go on the fix I think it is
better to allow the empty `title` object and have the backend handle it
gracefully, rather than add code to the Javascript in `create_chart_2`
to arbitrarily add empty `title` objects for some fields that don't
exist on the page and are not rendered anywhere.

I guess in the longer term (once all current charts are Chartbuilder 2
charts) we might think about dropping support for x and y axis labels
altogether. But they are currently included in cart download data, and
for now it feels like we need to support data both with and without axis
titles.