Add warning when not uploading object
 ## Summary
I spent quite a while trying to work out why a file wasn't being
uploaded to s3, despite the fact we were clearly calling `s3.write`.
This method silently refuses to write an object if it can't work out
what kind of file it is. It feels nice to add a message indicating that
the file is not being written because of this reason, to prevent any
wasted time trying to work out why it's not being uploaded.