Add black formatter for Python files
 ## Summary
This adds the 'black' Python code formatter to our repository and
includes a git hook that will enforce code formatting on commit so that
we don't need to worry about running it manually. The black formatter is
an opinionated, minimally-configurable code formatter that looks to
prevent discussions on personal preference when styling code by
enforcing well-known and widely-agreed-upon standards in formatting
Python code. The only option I've manually configured is line length,
which is now set to wrap at 120 characters.

The formatting decisions used by Black are documented on its github page
here: https://github.com/ambv/black

To install the hook, run `./scripts/install-hooks.sh`. The hook will
fail if the black formatter is not installed. It can be installed with
`pip install -r requirements-test.txt`. It's also likely that you can
configure your IDE to use `black` as its default code formatter, which
may be preferable.