Fix bug in topic table on homepage
 ## Summary
The homepage topic table was only designed to display 4-6 topics. Any
extra topics retrieved from database are all added to the same (2nd) row
on the page, which displays in a broken way.

This refactors the homepage to be able to display an arbitrary number of
topics and adds a test which asserts the display is correct.