Add migration for population topic
 ## Summary
Adds the population topic to the site via a data migration. We have a
weird conditional in the migration to allow this to be run against the
test database, which is generally empty, as our new topic has a
dependency on a 'homepage' record being in the database in order to set
up the correct page structure.

 ## Ticket
https://trello.com/c/AhgrGiIL/857