Add custom styling for results
 ## Summary
Override some of the styling injected by Google to make it look more
in-keeping with the rest of our site. This mostly just removes some
padding and sets a max-width for text. This is not done in the most
loving way as we are using Google Custom Search to prototype our search,
so expect to revisit this to either remove it (if search is not used or
not needed), or implement it more fully/properly if it's a needed
feature.