Refactor topic page and fix access bug
 ## Summary
Fixes a typo in `CREATE_MEASURE` (spelled as `CREATE_MASURE`) meant
that the link to create a new measure was not ever displayed to users
if a topic did not already contain measures.

Refactors the `topic.html` page to reduce duplication around this bit of
code as well, as part of the problem was that we were checking this
permission in two places - once if there were already measures, and once
if there are not already measures - and presumably we were only really
ever dealing with the first situation, so never noticed the bug in the
second. By removing the duplication we will make it immediately obvious
if a regression appears here in the future.