Refactor grouping and template usage
 ## Summary
Removes python inbuilt grouping in favour of default dictionaries to
build the datastructure of years->months->pages to pass into the
template.

Python3.6+ dictionaries preserve order based on insertion of keys, but
we don't specifically rely on this as we also perform ordering within
the template to ensure we display page updates in reverse chronological
order.