Allow google analytics site-wide
 ## Summary
When adding search to the site, we had to open up our Content Security
Policy. In the refactor for this, we accidentally made google analytics
only available on the search site in commit
d9b4c9fdc9d8fe178370ea340570b15807c4556d. This returns the default CSP
to the old settings, which should give us proper analytics again.

 ## Ticket
https://trello.com/c/lSaOw2Dm/897