Limit pages/links shown on static site
 ## Summary
On the static site, we only want to display links to topics and
subtopics that contain published measures - i.e. pages that users can
actually visit. On the CMS, we need users to be able to see all topics
and subtopics, in case they want to add a new measure to a
topic/subtopic that doesn't currently have any.

This adds logic that hides/shows topic and subtopic links following the
above logic. This will make it easier to release the later data
migrations and redirects when we are moving measures between
subtopics/topics.

 ## Ticket
https://trello.com/c/ikAJhzHv/870