Migrate business measures
 ## Summary
Migrate existing business and self-employment measures from Work, pay
and benefits to Workforce diversity. Note that we do this slightly
differently than before as we are moving the entire subtopic between
topics, not just moving measures between subtopics.

We will only add redirects for the published measures under the original
topic+subtopic, however - not a full redirect for the subtopic.

 ## Ticket
https://trello.com/c/3KmAZOT7/871