Update to Gulp 4 to fix lodash security vulnerability
Github has started to warn us about a security vulnerability in old
versions of lodash, which is a dependency of several of the mpm
packages we use.

Upgrading to Gulp version 4 pulls in a newer version of lodash that is
patched for the security vulnerability.

The move from Gulp 3 to Gulp 4 is a breaking change, I've made the minimal changes to gulpfile.js here for it to continue to work.