Fix url generation for social links
 ## Summary
Our social links on production have two slashes between the domain and
the page fragment. This is because the RDU_SITE is configured on
production with a trailing slash, and the fragment generation by flask
includes a preceding slash.

We should make sure to strip slashes from the domain before appending
the url fragment. I've also updated the RDU_SITE variable in production
config to remove the trailing slash, so this is really just making it
doubly sure.