Accessibility and bug fixes
 ## Summary
Enables a more accessible navigation flow for v2 of ChartBuilder and
TableBuilder by using `button` elements instead of `div` elements and
cancelling the form submit function. This allows better keyboard and
voice navigation on the page.

Also includes a bug fix for the cancel button, which didn't work due to
mismatched IDs between the HTML and the JS.