Run gulp using local version of gulp
This fixes a bug I encountered whereby `gulp --production` was failing with:

```javascript
TypeError: Cannot read property 'apply' of undefined
    at /usr/local/lib/node_modules/gulp/bin/gulp.js:129:20
    at process._tickCallback (internal/process/next_tick.js:61:11)
    at Function.Module.runMain (internal/modules/cjs/loader.js:746:11)
    at startup (internal/bootstrap/node.js:238:19)
    at bootstrapNodeJSCore (internal/bootstrap/node.js:572:3)
Frontend asset build check failed
```

This is, I think, because running `gulp --production` uses the version of gulp accessible in your PATH, which isn't necessarily the version installed in `node_modules` (ie the one specified by the `package-lock.json`).

`npx` is a package installed by `npm` which, as I understand it, makes sure that the version installed in `node_modules` is used. (Similar to `bundle exec` in Ruby land or `virtualenv` in Python land)