Don't require config vars that aren't used anywhere by the app
HARMONISER_ENABLED was dropped over a year ago in the code so that
harmoniser is always on, but remains in config until now.

The concept of multiple PUBLICATION_STATES was only needed prior to
initial launch of the website and was removed from codebase back in
April but remains in config until now.