Resize table/chart header inputs
 ## Summary
At the moment we display Chartbuilder/Tablebuilder 2 as text input
fields which are limited to one line. If you have a title that's a bit
longer, you can't see the end of it without clicking into the box and
side-scrolling.

The inputs are now textareas that automatically resize through JS. The
return key is also prevented to ensure that titles can't go onto two
lines. These are controlled through classes on the elements that the JS
looks for on page load (`js-autosize` and `js-prevent-return-key`).

 ## Ticket
https://trello.com/c/mLl9oW5H/950