Upload static assets to correct subdirectory
 ## Summary
A patch intended to upload static assets before deploying the rest of
the site, to prevent misconfigured css on the site during static builds,
accidentally uploaded the files to S3 with the wrong key - i.e. at the
root of the bucket rather than with the `static/` prefix. This commit
fixes that bug.

 ## Ticket
https://trello.com/c/KOFNKrtm/969