Don't attempt to render arbitrary files on the server
It seems better to only try to render a template if we know it's a file
we want to render, rather than anything that a user types in to the
address bar.

These pages are mostly going away soon when the new measures are
published to replace them, except (I think) ethnic-groups-and-data-collected.

Once we're down to one page we can make this prettier, but for now let's
patch it.