Fix table ordering element IDs
 ## Summary
We have fields for users to declare which column in their dataset
provides ordering information. We use jQuery to get the value of these
fields, but jQuery silently absorbs errors and returns a null/empty
value, which made us miss the fact that we had the wrong element ID in
the selector. This meant that ordering was never factored into the
table.

 ## Ticket
https://trello.com/c/r0rae1EX/1021