Update measure trello board columns
 ## Summary
The current single column, 'Received, in progress' currently leaves
measures looking like they're in progress for a lot longer than they
actually are. We're going to split this one column into two: 'Data
received' and 'Data checks in progress', to more clearly delineate
between the states where we've just received data from the department
and it's waiting to be prioritised/picked up, and when we're actually
actively doing something with it.

 ## Ticket
https://trello.com/c/5bv6FIsd/930