Fix attachment scanner integration
 ## Summary
We had code to make use of an integration from Heroku for scanning
uploaded documents, but a number of bugs in the implementation meant
this wouldn't ever work (although we also didn't have the correct
environment variable set on our instances to turn the functionality on -
ATTACHMENT_SCANNER_ENABLED).

This patch fixes up the scanning code, separates it our into a distinct
scanning service, and adds tests around the functionality. Once this has
shipped out, we need to enable the feature through the environment
variable above.

It also creates a parent exception for upload errors (UploadError) and
moves the existing upload errors into subclasses of that parent.

One noteworthy decision made in this: if our file scanner reports
that the file scan is 'pending', we will still let it be uploaded. We'll
emit a warning to the logger, so we could have sentry/papertrail alert
us to this case, but the alternative is to block the upload altogether.
I'm not sure how often scans will come back as pending, so I opted to be
permissive here. Maybe we shouldn't be?

Given that the attachment scanner calls out to an external service, I've
also mocked out the requests library across our entire test suite. Any
requests made will now raise `UnmockedRequestException` unless the
specific URL/response is mocked out subsequently. This should give us a
good level of confidence that our tests aren't sending fake/crap data
out to our providers.

We don't seem to have any direct unit tests for our `UploadService`
class, but that's too big a beast to handle as part of this ticket so
I'll be making a separate story for it on our tech debt board. I wanted
to at least do an integration test between the UploadService and the
ScannerService, but that proved a bit painful as the method also uses
a FileService that is set up differently and didn't seem easy to mock
out. We'll have to tackle this later.

We also didn't have an attachment scanner on our staging instance;
I've added one to it now.

 ## Ticket
https://trello.com/c/nT2ysJkS/963