Add latest measure title as page context
 ## Summary
When looking at all versions of a measure, the page doesn't show the
name of the measure anywhere, making it easy to lose context if you stop
paying attention or get distracted briefly.

This adds the name of the most recent version of the measure to the page
title as a secondary/context heading, including text to make it read
nicely for screen readers.

 ## Ticket
https://trello.com/c/cOUCruoT/778