Change error message on uploading bad file
 ## Summary
**Note**: This PR sits on top of the #747; please only review final
commit.

If a user uploads a file and we can't detect any encodings for that
file, we send back a vague error message about not being able to detect
an encoding. It's more helpful for the user to tell them that they ought
to be uploading a CSV file. We have a separate error message already
available if they upload a CSV/text file that has an unsupported
encoding.

Also adds a couple of tests around the `validate_file` method of the
`UploadService` - note that the rest of the class is untested. We have a
tech improvement ticket for addressing this at a later date.

 ## Ticket
https://trello.com/c/MjyJbEWG/668