Be really sure we have an error code to render
We have recently seen an error in production where the render_error
method was attempting to render the template `None.html`.

So it looks like the underlying error did have an attribute "code" but
it's value was None, so the default 500 was not being set.

The change here should ensure that we hit the 500.html template for any
future errors that come through with `code = None`.