Add management command for updating error pages
 ## Summary
This adds a management command to the app which can be used to update
error pages when the base template is changed. Includes a minor refactor
of the existing site generation code to allow for creating a temporary
build directory to use for storing files before upload.

This also brings in the two pre-existing static pages we have in our
errors bucket right now for Heroku to serve if it fails to get a
response from upstream (CloudFront).

The directory structure of `static_site/errors` has changed slightly to
reflect the structure of the S3 bucket as it stands now. This has Heroku
errors at the root level, and static site (http status code) errors
inside a `major-errors` folder. This is slightly unusual, but the
easiest change to make right now to keep everything compatible.
CloudFront loads errors from `major-errors`, Heroku from the root
directory. At the moment we also have a `403.html` in the root
directory, but I think this is unused. I'm not going to create a copy in
this repository. I can't see anything that serves the 403 from there, so
we should be able to delete it eventually.

 ## Ticket
https://trello.com/c/JvoRkcFl/983