Set mimetype for .map files so they will be uploaded to S3
The mimetype isn't auto-detceted for these files so up until now they
have been excluded from the push up to S3 and the build process
complains about them.

Fixing these will make it easier to debug the live website.

The received wisdom seems to be that the correct mimetype is
"application/json" - see:
https://stackoverflow.com/questions/19911929/what-mime-type-should-i-use-for-javascript-source-map-files
https://stackoverflow.com/questions/18809567/what-is-the-correct-mime-type-for-min-map-javascript-source-files