Always uglify assets
This remove the `--production` flag from gulp, which was used to decide whether to 'uglify' the CSS and javascript assets or not.

The main justification for this is that the uglify process can fail if the syntax is invalid, and so it's more useful to be notified of this failure early, rather than waiting until the tests run.

There's also no reason not to serve uglified assets in local development now, as sourcemaps can be used by modern browsers to resolve uglified and minified javascript back to its original form for debugging.