Move main element to base template
 ## Summary
This patch moves the definition of the `main` element to the base
template, so sub-templates no longer need to define this as part of
their `main_content` block. For templates that need to override the
definition of the main block, for example to add additional attributes
or override the class/role/etc, the block `main_element_definition` has
been added.

 ## Ticket
https://trello.com/c/zBytRdtM/1042