Centralise breadcrumbs structure
 ## Summary
Rationalises the way that our templates declare their breadcrumbs into a
variable, `breadcrumbs`, defined near the top of each sub-template
(specifically, after imports and before any block definitions). You
define the breadcrumbs as a list of dicts with the `url` and `text`
keys. An optional `event` key can be provided to declare when you want to
include a Google Analytics event on link clicks. These should probably
be added for all breadcrumb links, to be consistent, or removed if this
functionality isn't actually useful - but this decision is left for the
future (has been recorded in a tech improvement ticket for now).

 ## Ticket
https://trello.com/c/zBytRdtM/1042