Use block names from GOV.UK Design System
 ## Summary
This brings us in line with the current block names (and naming
convention of pascalCase) from the GOV.UK Design System. Some blocks
have been renamed, and some extra ones have been added to the base
template, to provide reasonable hooks and overrides for the structure.

All of the blocks that we export, as well as the structure and purpose,
has been documented in a README.md file at the root of the `templates`
directory.

This retires the `scripts` block in favour of the Design System's
`bodyEnd` block. Derived templates have been updated to use the
`bodyEnd` block. One side-effect of this change is that, because the
`bodyEnd` block contains content in the base template, all derived
templates need to have a Jinja2 call to include the parent's block
content in the derived template via `{{ super() }}`.

 ## Ticket
https://trello.com/c/zBytRdtM/1042