Remove CMS base template
 ## Summary
We have a derived base template for the CMS where the only difference is
the inclusion of CMS CSS. The JavaScript is included in a conditional
check in the root (static) base template, and doing this with the CSS
seems safer and would be more consistent. This patch therefore removes
the CMS base template and includes the CMS CSS in the base template, if
as long as it isn't being rendered in static mode.

 ## Ticket
https://trello.com/c/zBytRdtM/1042