Consolidate flash message handling
 ## Summary
We currently have template-specific handling of flash messages, which
means that we can only handle flash messages on specific pages of the
site - and we handle them in different ways in different places.

This moves flash message handling to the base template in a new block
defined for this purpose. It can be overridden by pages which don't meet
the standard (i.e. most common) way that we handle flash messages until
we can fix up the views that work in a unique way.

This reduces the amount of template code we have, and makes it less
likely that we will have differences in how we render flashes depending
on the specific page the user is on. It also means that flash messages
can be shown to the user regardless of which page they end up on.

 ## Ticket
https://trello.com/c/zBytRdtM/1042