Fix chart/table title height on new dimensions
 ## Summary
When creating a new chart/table, the box for entering the title started
with a bad height (too small). This fix changes when the event for
resizing is fired by chart/table builder to every time the table is
regenerated, rather than just the first time the page is setup (which
does nothing in the case of new charts/tables.

 ## Ticket
https://trello.com/c/q2ypmIGE/1070