Update code "11A" to "12A"
There are 12 things in this list, so 12A is a better code for it.

Making this change will mean that the current charts and tables that use
this "preset" will not reload the "Ethnicity" in the chart/table builder
editor page.  But the preset will be listed in the drop-down and people
will just need to select it again.

This will only affect four (4) current draft measures, so doesn't feel
too bad an impact on users, and it will be nice to get our codes in
order.