Make sync_classification a data migration not managed by Alembic
There are no structural changes in this migration so we run it as a one-
off data migration rather than in the Alembic migrations.

I have removed the final
`UPDATE categorisation SET code = '11A' WHERE code = '12A';`
as the previous commit changes the code from 11A to 12A in the
classifications csv, so the existing "12A" are correct.