BUGFIX grouped bar chart upgrade v1 to v2
The chartbuilder v1 to v2 upgrade currently expects v1 grouped bar
charts to have a list of objects as series data.  However, a small
number of grouped bar charts have a list of numbers (int or float), and
wen the upgrade is attempted on these charts it fails with a 500 error.

The additional check here allows these "number-list" v1 charts to upgrade
to v2 charts properly.