Remove metadata from CSV files
This removes the metadata (eg page title, date, URL)  which is currently injected into the CSV files.

Research showed that this metadata was unhelpul, as many tools which read CSV assume that the first line either contains headers or the first line of data, and so was having to be manually removed.

The code also sets the MIME type of the CSV files, when served via the Flask app, to correctly be `text/csv` rather than `text/plain`.