Transposed cells in tables
A bug with building tables was discovered recently by content and
analysts, wherein the live preview of a table was showing values in the
correct order, by the 'static' preview (via preview the measure) was
displaying some values transposed in an unknown order.

This happened for grouped tables where the source data was not provided
in a consistent format. Specifically, where the subgroups were not in
the same order, as per the example below. The table also needed to have
no explicit row ordering set (explicit row ordering prevented the bug).

| Ethnicity | Category |
|-----------|----------|
| White     | One      |
| White     | Two      |
| Black     | Two      |
| Black     | One      |

Some steps in the table builder process look at the order of the
subgroup (here, `Category`), and then extract further values based on
the assumption that all other major groups (here, `Ethnicity`) have a
consistent sub-group ordering.

This was a flawed assumption for the dataset in question today. This
patch applies a default alphabetical row ordering if no order is
explicitly set.

 ## Ticket
https://trello.com/c/mrtsskDU/1103