Update tablebuilder2 for row/column ordering
 ## Summary
We need users to explicitly select a row/column ordering for tables they
build, otherwise there's a risk that some cells will be transposed if
the data for the table is provided in an inconsistent format.

This patch removes the 'None' ordering option for tables, and sets a
reasonable default in cases where it hasn't previously been provided.

 ## Ticket
https://trello.com/c/mrtsskDU/1103